checks if already installed before download or install

+++ install/desktop/app-chrome.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 # Browse the web with the most popular browser. See https://www.google.com/chrome/
-cd /tmp
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo apt install -y ./google-chrome-stable_current_amd64.deb
-rm google-chrome-stable_current_amd64.deb
-cd -
+if ! command -v google-chrome &> /dev/null; then
+  cd /tmp
+  wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+  sudo apt install -y ./google-chrome-stable_current_amd64.deb
+  rm google-chrome-stable_current_amd64.deb
+  cd -
+fi
