+++ install/desktop/app-localsend.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-cd /tmp
-LOCALSEND_VERSION=$(curl -s "https://api.github.com/repos/localsend/localsend/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
-wget -O localsend.deb "https://github.com/localsend/localsend/releases/latest/download/LocalSend-${LOCALSEND_VERSION}-linux-x86-64.deb"
-sudo apt install -y ./localsend.deb
-rm localsend.deb
-cd -
+if ! command -v localsend_app &> /dev/null; then
+  cd /tmp
+  LOCALSEND_VERSION=$(curl -s "https://api.github.com/repos/localsend/localsend/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
+  wget -O localsend.deb "https://github.com/localsend/localsend/releases/latest/download/LocalSend-${LOCALSEND_VERSION}-linux-x86-64.deb"
+  sudo apt install -y ./localsend.deb
+  rm localsend.deb
+  cd -
+fi
