+++ install/terminal/app-zellij.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
-cd /tmp
-wget -O zellij.tar.gz "https://github.com/zellij-org/zellij/releases/latest/download/zellij-x86_64-unknown-linux-musl.tar.gz"
-tar -xf zellij.tar.gz zellij
-sudo install zellij /usr/local/bin
-rm zellij.tar.gz zellij
-cd -
+if ! command -v zellij &> /dev/null; then
+  cd /tmp
+  wget -O zellij.tar.gz "https://github.com/zellij-org/zellij/releases/latest/download/zellij-x86_64-unknown-linux-musl.tar.gz"
+  tar -xf zellij.tar.gz zellij
+  sudo install zellij /usr/local/bin
+  rm zellij.tar.gz zellij
+  cd -
+fi
 
 mkdir -p ~/.config/zellij/themes
 [ ! -f "$HOME/.config/zellij/config.kdl" ] && cp ~/.local/share/omakub/configs/zellij.kdl ~/.config/zellij/config.kdl
